chore(antiddos): move antiddos data source to deprecated
